fixed bug with force_update

+++ library/vmware_drs.py
@@ -360,7 +360,8 @@ class VMWareDRS(object):
     def update(self):
         """Update VMWare DRS rule."""
         updated = False
-        self.delete()
+        if self.rule_exists:
+            self.delete()
         self.create()
 
         if self.check():
